Updated headers

+++ src/backend/include/graph.hpp
@@ -6,6 +6,7 @@
 #include "settings.hpp"
 #include "tasks.hpp"
 #include <array>
+#include <cstdint>
 #include <string>
 #include <unordered_map>
 #include <vector>
@@ -22,7 +23,7 @@ public:
   std::vector<std::string> task_names;
   std::vector<std::vector<taskid_t>> read_data;
   std::vector<std::vector<taskid_t>> write_data;
-  std::vector<std::array<std::array<int, 4>, num_device_types>> variant_info;
+  std::vector<std::array<std::array<int64_t, 4>, num_device_types>> variant_info;
   std::vector<std::vector<taskid_t>> dependencies;
 
   std::unordered_map<std::string, taskid_t> task_name_to_id;
@@ -60,7 +61,7 @@ public:
     read_data.push_back(std::vector<taskid_t>());
     write_data.push_back(std::vector<taskid_t>());
 
-    std::array<std::array<int, 4>, num_device_types> zero_variant_array{};
+    std::array<std::array<int64_t, 4>, num_device_types> zero_variant_array{};
     variant_info.push_back(zero_variant_array);
     dependencies.push_back(std::vector<taskid_t>());
     return task_id;
@@ -248,7 +249,7 @@ public:
         add_write_data(j + offset, write_data[j]);
 
         for (std::size_t k = 0; k < num_device_types; k++) {
-          std::array<int, 4> info = variant_info[j][k];
+          std::array<int64_t, 4> info = variant_info[j][k];
           variant_info[j + offset][k] = info;
         }
       }
@@ -272,7 +273,7 @@ public:
         add_write_data(i + offset, graph.write_data[i]);
 
         for (std::size_t j = 0; j < num_device_types; j++) {
-          std::array<int, 4> info = graph.variant_info[i][j];
+          std::array<int64_t, 4> info = graph.variant_info[i][j];
           variant_info[i + offset][j] = info;
         }
       }
@@ -290,7 +291,7 @@ public:
       tasks.set_write(i, get_write_data(i));
       tasks.set_type(i, get_type(i));
       for (std::size_t j = 0; j < num_device_types; j++) {
-        std::vector<int> info = get_variant_info(i, static_cast<DeviceType>(j));
+        std::vector<int64_t> info = get_variant_info(i, static_cast<DeviceType>(j));
         if (info[0] == 1) {
           tasks.add_variant(i, static_cast<DeviceType>(j), info[1], info[2], info[3]);
         }
